+++ Assets/_package_/_main_/Editor/Develop/PackagePathGenerator.cs
@@ -4,7 +4,7 @@ using System.CodeDom;
 
 namespace UPMTool
 {
-    public class PackagePathGenerator
+    public static class PackagePathGenerator
     {
         public static void Generate(string nameSpace, string path, string className = "PackagePath")
         {