+++ Assets/_package_/_main_/Editor/UPMToolExtension/UPMToolExtension.cs
@@ -196,6 +196,7 @@ namespace UPMToolDevelop
 
         public void OnPackageAddedOrUpdated(PackageInfo packageInfo)
         {
+            
         }
 
         public void OnPackageRemoved(PackageInfo packageInfo)